make RxJava helpers optional in Updater
